requirements.txt
streamlit 
opencv
numpy
matplotlib.pyplot